Feat: 설정페이지 게시판 카테고리 삭제
