+++ src/routes/Setting/Board/index.tsx
@@ -11,8 +11,6 @@ import {
 } from "./../../../states/board";
 import {
   CategoryDeleteButton,
-  CategoryDeleteContent,
-  CategoryDeleteSpan,
   CategoryListBox,
   SettingButtonBox,
   SettingCancelButton,
@@ -25,21 +23,18 @@ import {
 } from "./style";
 import { useRecoilState, useRecoilValue } from "recoil";
 import { useBoardCategoryUpdate } from "../../../hooks/board/category/useBoardCategoryUpdate";
-import { useBoardCategoryDelete } from "../../../hooks/board/category/useBoardCategoryDelete";
 
 const BoardManagement = () => {
   const [boardName, setBoardName] = useState<string>("");
   const { boardCategoryList } = useBoardCategoryList();
 
   const [isOpenCreateModal, setIsOpenCreateModal] = useState(false);
-  const [isOpenDeleteModal, setIsOpenDeleteModal] = useState(false);
 
   const categoryId = useRecoilValue(settingCategoryId);
   const [categoryName, setCategoryName] = useRecoilState(settingCategoryName);
 
   const createMutation = useBoardCategoryCreate(boardName);
   const categoryUpdate = useBoardCategoryUpdate(categoryId, categoryName);
-  const categoryDelete = useBoardCategoryDelete();
 
   const handleCreateCategory = () => {
     createMutation.mutate();
@@ -47,19 +42,10 @@ const BoardManagement = () => {
     setIsOpenCreateModal(false);
   };
 
-  const handleDeleteCategory = () => {
-    categoryDelete.mutate(categoryId);
-    setIsOpenDeleteModal(false);
-  };
-
   const handleClickCreateModal = useCallback(() => {
     setIsOpenCreateModal(!isOpenCreateModal);
     setBoardName("");
   }, [isOpenCreateModal]);
-
-  const handleClickDeleteModal = useCallback(() => {
-    setIsOpenDeleteModal(!isOpenDeleteModal);
-  }, [isOpenDeleteModal]);
 
   return (
     <>
@@ -125,30 +111,6 @@ const BoardManagement = () => {
             <SettingSaveButton
               style={FONT.SUBTITLE2}
               onClick={handleCreateCategory}
-            >
-              확인
-            </SettingSaveButton>
-          </SettingButtonBox>
-        </Modal>
-      )}
-
-      {isOpenDeleteModal && (
-        <Modal onClickToggleModal={handleClickDeleteModal} size={"small"}>
-          <SettingTitle style={FONT.HEADING}>게시판 삭제</SettingTitle>
-          <CategoryDeleteContent style={FONT.SUBTITLE2}>
-            <CategoryDeleteSpan>{categoryName}</CategoryDeleteSpan> 게시판을
-            삭제하시겠습니까?
-          </CategoryDeleteContent>
-          <SettingButtonBox>
-            <SettingCancelButton
-              style={FONT.SUBTITLE2}
-              onClick={handleClickDeleteModal}
-            >
-              취소
-            </SettingCancelButton>
-            <SettingSaveButton
-              style={FONT.SUBTITLE2}
-              onClick={handleDeleteCategory}
             >
               확인
             </SettingSaveButton>
